+++ packages/powersync/lib/sqlite3_common.dart
@@ -1,4 +1,4 @@
-/// Re-exports [sqlite3](https://pub.dev/packages/sqlite3) to expose sqlite3 without
+/// Re-exports [sqlite3_common](https://pub.dev/packages/sqlite3) to expose sqlite3_common without
 /// adding it as a direct dependency.
 library;