+++ packages/powersync/lib/src/open_factory/open_factory_stub.dart
@@ -18,4 +18,9 @@ class PowerSyncOpenFactory extends open_factory.AbstractPowerSyncOpenFactory {
   void setupFunctions(CommonDatabase db) {
     throw UnimplementedError();
   }
+
+  @override
+  String getLibraryForPlatform({String? path}) {
+    throw UnimplementedError();
+  }
 }