+++ packages/powersync_flutter_libs/linux/powersync_flutter_libs_plugin.cc
@@ -1,0 +1,7 @@
+#include "include/powersync_flutter_libs/powersync_flutter_libs_plugin.h"
+
+#include <flutter_linux/flutter_linux.h>
+
+void powersync_flutter_libs_plugin_register_with_registrar(FlPluginRegistrar *registrar) {
+  
+}